apps/api/requirements.txt
## Message Queue
kafka-python==2.0.2
six
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Databases & Caching
redis==5.0.1
psycopg2-binary==2.9.9
clickhouse-connect>=0.6.0
boto3==1.29.7
minio==7.2.0

# Schema Registry & DLQ (HA Features)
confluent-kafka[avro]==2.3.0
avro-python3==1.10.2
fastavro==1.9.4

# GraphQL
strawberry-graphql==0.219.0

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
python-dotenv==1.0.0
PyYAML==6.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
